+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## UNRELEASED
+
+### Fixed
+
+- Symfony uploaded file moving (`FixSymfonyFileMovingListener` was added for this) [#43]
+
+[#43]:https://github.com/spiral/roadrunner-laravel/issues/43
+
 ## v5.0.0
 
 ### Added